Redone stories
stories now are in the center, with the larger ones taking up more space, and blake is now added to the first story. The links at the bottom now exist, and are similar in style to the ones at the top. The footer now has the dots to make it look better. Texts were centered.