Mostly done
the page now is done in terms of content, and looks like it should for the final. The entire page now has a background, the navs now have a border, the stories have a border, and the last few links have been added. The last nav has had its header changed, so that the navs aren't pushed to the footer.